Extract collection to Collectible module
